requirements.txt
pytest
selene==2.0.0rc9
requests~=2.32.3
allure-pytest
allure-python-commons~=2.14.1
jsonschema~=4.24.0